adding the search category UI
