+++ app/src/main/java/com/example/clicker/presentation/stream/HorizontalChat.kt
@@ -25,9 +25,9 @@ import com.example.clicker.presentation.stream.views.TestingNewBottomModal
 
 import com.example.clicker.presentation.stream.views.chat.chatSettings.ChatSettingsColumn
 import com.example.clicker.presentation.stream.views.chat.ChatUI
-import com.example.clicker.presentation.stream.views.chat.chatSettings.ChannelInfoLazyColumn
 import com.example.clicker.presentation.stream.views.chat.chatSettings.ChatSettingsViewModel
 import com.example.clicker.presentation.stream.views.dialogs.WarningDialog
+import com.example.clicker.presentation.streamInfo.ChannelInfoLazyColumn
 import com.example.clicker.presentation.streamInfo.ContentClassificationCheckBox
 import com.example.clicker.presentation.streamInfo.StreamInfoViewModel
 
@@ -80,6 +80,9 @@ fun HorizontalChat(
         scope.launch {
             editChannelInformationModalState.hide()
         }
+    } }
+    val removeCategory:() -> Unit = remember(streamInfoViewModel) { {
+        streamInfoViewModel.removeCategory()
     } }
 
 
@@ -217,7 +220,8 @@ fun HorizontalChat(
                 checkedBrandedContent = streamInfoViewModel.brandedContent.value,
                 changeBrandedContent={newValue ->changeBrandedContent(newValue)},
                 categoryResponse = streamInfoViewModel.gameCategoryResponse.value,
-                refreshChannelInformation ={refreshChannelInformation()}
+                refreshChannelInformation ={refreshChannelInformation()},
+                removeCategory={removeCategory()}
             )
         }
     ) {
