+++ app/src/main/java/com/example/clicker/presentation/stream/views/chat/chatSettings/ChatSettings.kt
@@ -177,727 +177,6 @@ fun ChatSettingsColumn(
 
 }
 
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun ChannelInfoLazyColumn(
-    streamTitle:String,
-    changeStreamTitle:(String)->Unit,
-    titleLength:Int,
-
-    tagList:List<String>,
-    tagTitle:String,
-    tagTitleLength:Int,
-    changeTagTitle:(String)->Unit,
-    addTag:(String)->Unit,
-    removeTag:(String)->Unit,
-
-    contentClassificationCheckBox: ContentClassificationCheckBox,
-    changeContentClassification:(ContentClassificationCheckBox)->Unit,
-
-    selectedLanguage:String?,
-    changeSelectedLanguage:(String)->Unit,
-    closeChannelInfoModal:()->Unit,
-
-    checkedBrandedContent:Boolean,
-    changeBrandedContent:(Boolean)->Unit,
-
-    categoryResponse: Response<Game>,
-    refreshChannelInformation:()->Unit,
-){
-
-
-    LazyColumn(modifier = Modifier
-        .fillMaxSize()
-        .background(MaterialTheme.colorScheme.primary)) {
-        stickyHeader {
-            SaveCloseChannelInformationHeader(
-                closeModal={closeChannelInfoModal()}
-            )
-        }
-
-        item{
-            ChannelInfoTitle(
-                streamTitle=streamTitle,
-                changeStreamTitle={newValue ->changeStreamTitle(newValue)},
-                titleLength=titleLength
-            )
-        }
-        item {
-            GameCategory(
-                categoryResponse=categoryResponse,
-                refreshChannelInformation= { refreshChannelInformation() }
-            )
-        }
-
-        item{
-            ChannelTagsInfo(
-                tagList =tagList,
-                tagTitle=tagTitle,
-                tagTitleLength=tagTitleLength,
-                changeTagTitle={newValue ->changeTagTitle(newValue)},
-                addTag={newTag ->addTag(newTag)},
-                removeTag={oldTag -> removeTag(oldTag)}
-            )
-
-        }
-        item{
-            Spacer(modifier =Modifier.height(20.dp))
-        }
-
-        item{
-            ContentClassificationBox(
-                contentClassificationCheckBox=contentClassificationCheckBox,
-                changeContentClassification={newClassification ->changeContentClassification(newClassification)}
-            )
-        }
-        item{
-            Spacer(modifier =Modifier.height(20.dp))
-        }
-
-        item{
-            StreamLanguage(
-                selectedLanguage =selectedLanguage,
-                changeSelectedLanguage = {newValue ->
-                    changeSelectedLanguage(newValue)
-                }
-            )
-        }
-        item{
-            Spacer(modifier =Modifier.height(20.dp))
-        }
-        item{
-            BrandedContentInfo(
-                checkedBrandedContent=checkedBrandedContent,
-                changedCheckedBrandedContent={newValue -> changeBrandedContent(newValue)}
-            )
-        }
-
-
-        item{
-            Spacer(modifier =Modifier.height(10.dp))
-        }
-        item{
-            ShareButton()
-        }
-
-    }
-}
-
-@Composable
-fun GameCategory(
-    categoryResponse: Response<Game>,
-    refreshChannelInformation:()->Unit,
-){
-    Column(
-        modifier = Modifier.padding(10.dp)
-    ) {
-        Text("Category",color = MaterialTheme.colorScheme.onPrimary,
-            fontSize = MaterialTheme.typography.headlineMedium.fontSize
-        )
-        Divider(thickness = 1.dp, color = MaterialTheme.colorScheme.secondary.copy(0.8f))
-        Spacer(modifier = Modifier.height(5.dp))
-        ElevatedCard(
-            elevation = CardDefaults.cardElevation(
-                defaultElevation = 6.dp
-            ),
-            modifier = Modifier.fillMaxWidth(),
-            shape =RoundedCornerShape(5.dp),
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(72.dp)
-            ){
-                when(val data = categoryResponse){
-                    is Response.Loading ->{
-                        CircularProgressIndicator(modifier = Modifier.align(Alignment.Center))
-                    }
-                    is Response.Success->{
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            verticalAlignment = Alignment.CenterVertically
-                        ){
-                            SubcomposeAsyncImage(
-                                modifier = Modifier
-                                    .height(72.dp)
-                                    .width(52.dp)
-                                    .clip(RoundedCornerShape(5.dp)),
-                                model = data.data.box_art_url,
-                                loading = {
-                                    Column(modifier = Modifier
-                                        .height(72.dp)
-                                        .width(52.dp)
-                                        .background(MaterialTheme.colorScheme.primary),
-                                        verticalArrangement = Arrangement.Center,
-                                        horizontalAlignment = Alignment.CenterHorizontally
-                                    ){
-                                        CircularProgressIndicator()
-                                    }
-                                },
-                                contentDescription ="Loading thumbnail of stream game"
-                            )
-
-                            Spacer(modifier =Modifier.width(10.dp))
-                            Text(data.data.name,
-                                color = Color.White,
-                                fontSize = MaterialTheme.typography.headlineMedium.fontSize,maxLines = 1, overflow = TextOverflow.Ellipsis)
-                        }
-                        Icon(
-                            painter = painterResource(id = R.drawable.baseline_close_24),
-                            contentDescription ="delete category" ,
-                            modifier = Modifier
-                                .align(Alignment.TopEnd)
-                                .padding(10.dp),
-                            tint = Color.White
-                        )
-                    }
-                    is Response.Failure ->{
-                        RefreshCategoryButton(
-                            refreshChannelInformation={refreshChannelInformation()}
-                        )
-                    }
-
-            }
-
-
-
-
-            }
-
-
-        }
-
-    }
-}
-
-@Composable
-fun RefreshCategoryButton(
-    refreshChannelInformation:()->Unit,
-){
-    Column(modifier = Modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ){
-        Text("Failed to get Category",color = Color.Red)
-        Button(
-            onClick={},
-            shape = RoundedCornerShape(5.dp),
-            colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.secondary)
-        ){
-
-            Text("Refresh", color = MaterialTheme.colorScheme.onSecondary)
-        }
-        Spacer(modifier =Modifier.height(10.dp))
-    }
-}
-
-@Composable
-fun BrandedContentInfo(
-    checkedBrandedContent:Boolean,
-    changedCheckedBrandedContent:(Boolean)->Unit,
-){
-    Column(
-        modifier = Modifier.padding(10.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ){
-            Text("Branded Content",color = MaterialTheme.colorScheme.onPrimary,
-                fontSize = MaterialTheme.typography.headlineMedium.fontSize
-            )
-            CustomCheckBox(
-                checked=checkedBrandedContent,
-                changeChecked={changedCheckedBrandedContent(!checkedBrandedContent)}
-            )
-        }
-        Divider(thickness = 1.dp, color = MaterialTheme.colorScheme.secondary.copy(0.8f))
-        Text("Let viewers know if your stream features branded content. This includes paid product placement, endorsement, or" +
-                " other commercial relationships", color = MaterialTheme.colorScheme.onPrimary, fontSize = MaterialTheme.typography.headlineSmall.fontSize)
-    }
-}
-
-@Composable
-fun SaveCloseChannelInformationHeader(
-    closeModal:()->Unit
-){
-    Row(modifier = Modifier
-        .fillMaxWidth()
-        .background(MaterialTheme.colorScheme.primary)
-        .padding(10.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ){
-        Icon(painter = painterResource(id =R.drawable.baseline_close_24),
-            contentDescription ="close" ,
-            tint = MaterialTheme.colorScheme.onPrimary,
-            modifier = Modifier.clickable {
-                closeModal()
-            }
-        )
-        Button(
-            onClick={},
-            shape = RoundedCornerShape(5.dp),
-            colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.secondary)
-        ){
-            Text("Save", color = MaterialTheme.colorScheme.onSecondary)
-        }
-
-    }
-}
-@Composable
-fun ShareButton(){
-
-    Box(modifier = Modifier
-        .fillMaxWidth()
-        .padding(10.dp)){
-        Button(
-            modifier = Modifier.fillMaxWidth(),
-            onClick={},
-            shape = RoundedCornerShape(5.dp),
-            colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.secondary)
-        ){
-            Row(modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 10.dp),
-                horizontalArrangement = Arrangement.SpaceBetween
-
-            ){
-                Icon(
-                    painter = painterResource(id = R.drawable.baseline_rocket_launch_24),
-                    contentDescription ="Share link to channel",
-                    tint = MaterialTheme.colorScheme.onSecondary
-                )
-                Text("Share link to channel" ,
-                    color = MaterialTheme.colorScheme.onSecondary,
-                    fontSize = MaterialTheme.typography.headlineMedium.fontSize
-                )
-                Icon(
-                    painter = painterResource(id = R.drawable.baseline_rocket_launch_24),
-                    contentDescription ="Share link to channel",
-                    tint = MaterialTheme.colorScheme.onSecondary
-                )
-            }
-        }
-    }
-
-
-}
-
-@Composable
-fun StreamLanguage(
-     selectedLanguage:String?,
-     changeSelectedLanguage:(String)->Unit,
-){
-    var expanded by remember { mutableStateOf(false) }
-    val languages = listOf("American Sign Language","Arabic","Bulgarian","Catalan","Chinese","Czech","Danish","Dutch","English","Finish","French","German",
-    "German","Greek","Hindi","Hungarian","Indonesian","Italian","Japanese","Korean","Malay","Norwegian","Polish","Portuguese","Romanian",
-    "Russian","Slovak","Spanish","Swedish","Tagalog","Thai","Turkish","Ukrainian","Vietnamese","Other")
-
-    Column(){
-
-        Text("Stream language",
-            fontSize = MaterialTheme.typography.headlineMedium.fontSize,
-            color = MaterialTheme.colorScheme.onPrimary,
-            modifier = Modifier.padding(start=10.dp)
-        )
-        Divider(thickness = 1.dp, color = MaterialTheme.colorScheme.secondary.copy(0.8f), modifier = Modifier.padding(horizontal=10.dp))
-    Box(modifier = Modifier.wrapContentSize(Alignment.BottomCenter)) {
-
-        OutlinedTextField(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 10.dp)
-                .clickable {
-                    expanded = true
-                },
-            enabled = false,
-            //todo: this is what is shown to the user as the selected choice
-            value = selectedLanguage ?: "Select language",
-            onValueChange = { },
-            label = { },
-            colors = TextFieldDefaults.colors(
-                disabledTextColor = Color.White,
-                disabledContainerColor = Color.DarkGray,
-                disabledTrailingIconColor = Color.Unspecified,
-                disabledLabelColor = Color.Unspecified,
-                disabledPlaceholderColor = Color.Unspecified,
-                disabledSupportingTextColor = Color.Unspecified,
-                disabledPrefixColor = Color.Unspecified,
-                disabledSuffixColor = Color.Unspecified
-            ),
-            trailingIcon = {
-                if (expanded) {
-                    Icon(
-                        painter = painterResource(id = R.drawable.baseline_keyboard_arrow_up_24),
-                        contentDescription = "Content Classification open"
-                    )
-                } else {
-                    Icon(
-                        painter = painterResource(id = R.drawable.keyboard_arrow_down_24),
-                        contentDescription = "Content Classification closed"
-                    )
-                }
-            }
-        )
-        DropdownMenu(
-            expanded = expanded,
-            onDismissRequest = { expanded = false },
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(
-                    Color.DarkGray
-                )
-                .padding(horizontal = 10.dp)
-        ) {
-            Box(modifier = Modifier.fillMaxWidth()) {
-                Column() {
-                    for (item in languages) {
-                        Spacer(modifier = Modifier.height(10.dp))
-                        Text(
-                            item,
-                            fontSize = MaterialTheme.typography.headlineMedium.fontSize,
-                            color = Color.White,
-                            modifier = Modifier
-                                .clickable {
-                                    changeSelectedLanguage(item)
-                                    expanded = false
-                                }
-                        )
-
-                    }
-                }
-
-
-                Icon(
-                    painter = painterResource(id = R.drawable.baseline_close_24),
-                    contentDescription = "Close language menu",
-                    modifier = Modifier
-                        .align(Alignment.TopEnd)
-                        .clickable {
-                            expanded = false
-                        }
-                )
-            }
-
-        }
-    }
-
-    }
-}
-
-@Composable
-fun ContentClassificationBox(
-    contentClassificationCheckBox: ContentClassificationCheckBox,
-    changeContentClassification:(ContentClassificationCheckBox)->Unit,
-) {
-    var expanded by remember { mutableStateOf(false) }
-
-    Column(){
-
-        Text("Content Classification",
-            fontSize = MaterialTheme.typography.headlineMedium.fontSize,
-            color = MaterialTheme.colorScheme.onPrimary,
-            modifier = Modifier.padding(start=10.dp)
-        )
-        Divider(thickness = 1.dp, color = MaterialTheme.colorScheme.secondary.copy(0.8f), modifier = Modifier.padding(horizontal=10.dp))
-        Box(modifier = Modifier.wrapContentSize(Alignment.BottomCenter)){
-            OutlinedTextField(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 10.dp)
-                    .clickable {
-                        expanded = true
-                    },
-                enabled = false,
-                //todo: this is what is shown to the user as the selected choice
-                value = "Content Classification",
-                onValueChange = { },
-                label = {  },
-                colors = TextFieldDefaults.colors(
-                    disabledTextColor = Color.White,
-                    disabledContainerColor = Color.DarkGray,
-                    disabledTrailingIconColor = Color.Unspecified,
-                    disabledLabelColor = Color.Unspecified,
-                    disabledPlaceholderColor = Color.Unspecified,
-                    disabledSupportingTextColor = Color.Unspecified,
-                    disabledPrefixColor = Color.Unspecified,
-                    disabledSuffixColor = Color.Unspecified
-                ),
-                trailingIcon = {
-                    if(expanded){
-                        Icon(painter = painterResource(id = R.drawable.baseline_keyboard_arrow_up_24), contentDescription ="Content Classification open" )
-                    }else{
-                        Icon(painter = painterResource(id = R.drawable.keyboard_arrow_down_24), contentDescription ="Content Classification closed" )
-                    }
-                }
-            )
-            DropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(
-                        Color.DarkGray
-                    )
-            ){
-                Box(){
-
-                    Column(){
-                        Spacer(modifier = Modifier.height(10.dp))
-                        ContentClassificationTextMenuItem(
-                            setExpanded={newValue -> //expanded=newValue
-                            },
-                            title = "Drugs, Intoxication, or Excessive Tobacco Use",
-                            selectText={},
-                            subtitle = "Excessive tobacco glorification or promotion, any marijuana consumption/use,legal drug and alcohol induced intoxication" +
-                                    ", discussions of illegal drugs",
-                            checked = contentClassificationCheckBox.drugsIntoxication,
-                            changedChecked = {checked ->
-                                val newClassification = contentClassificationCheckBox.copy(drugsIntoxication = checked)
-                                Log.d("Changingtheclassification","checked ->${checked}")
-                                Log.d("Changingtheclassification","newClassification ->${newClassification}")
-                                changeContentClassification(newClassification)
-                            }
-                        )
-                        Spacer(modifier = Modifier.height(10.dp))
-                        ContentClassificationTextMenuItem(
-                            setExpanded={newValue -> },
-                            title = "Gambling",
-                            selectText={},
-                            subtitle = "Participating in online or in-person gambling , poker or fantasy sports, that involve the exchange of real money",
-                            checked = contentClassificationCheckBox.gambling,
-                            changedChecked = {checked ->
-                                val newClassification = contentClassificationCheckBox.copy(gambling = checked)
-                                changeContentClassification(newClassification)
-                            }
-                        )
-                        Spacer(modifier = Modifier.height(10.dp))
-                        ContentClassificationTextMenuItem(
-                            setExpanded={newValue -> },
-                            title = "Significant Profanity or Vulgarity",
-                            selectText={},
-                            subtitle = "Prolonged, and repeated use of obscenities, profanities, and vulgarities, especially as a regular part" +
-                                    "of speech",
-                            checked = contentClassificationCheckBox.significantProfanity,
-                            changedChecked = {checked ->
-                                val newClassification = contentClassificationCheckBox.copy(significantProfanity = checked)
-                                changeContentClassification(newClassification)
-                            }
-
-                        )
-                        Spacer(modifier = Modifier.height(10.dp))
-                        ContentClassificationTextMenuItem(
-                            setExpanded={newValue -> },
-                            title = "Sexual Themes",
-                            selectText={},
-                            subtitle = "Content that focuses on sexualized physical attributes and activities, sexual topics, or experiences" ,
-                            checked = contentClassificationCheckBox.sexualThemes,
-                            changedChecked = {checked ->
-                                val newClassification = contentClassificationCheckBox.copy(sexualThemes = checked)
-                                changeContentClassification(newClassification)
-                            }
-                        )
-                        Spacer(modifier = Modifier.height(10.dp))
-                        ContentClassificationTextMenuItem(
-                            setExpanded={newValue -> },
-                            title = "Violent and Graphic depictions",
-                            selectText={},
-                            subtitle = "Simulations and/or depictions of realistic violence, gore, extreme injury or death",
-                            checked = contentClassificationCheckBox.violentGraphic,
-                            changedChecked = {checked ->
-                                val newClassification = contentClassificationCheckBox.copy(violentGraphic = checked)
-                                changeContentClassification(newClassification)
-                            }
-                        )
-                        Spacer(modifier = Modifier.height(10.dp))
-                        ContentClassificationTextMenuItem(
-                            setExpanded={newValue -> },
-                            title = "Mature-rated game",
-                            selectText={},
-                            subtitle = "Games that are rated Mature or less suitable for a younger audience",
-                            checked = contentClassificationCheckBox.matureRatedGame,
-                            changedChecked = {checked ->
-                                val newClassification = contentClassificationCheckBox.copy(matureRatedGame = checked)
-                                changeContentClassification(newClassification)
-                            }
-
-                        )
-                        Spacer(modifier = Modifier.height(10.dp))
-                    }
-                    Icon(
-                        painter = painterResource(id =R.drawable.baseline_close_24),
-                        contentDescription = "Close language menu",
-                        modifier = Modifier
-                            .align(Alignment.TopEnd)
-                            .clickable {
-                                expanded = false
-                            }
-                    )
-                }
-
-
-            }
-
-        }
-    }
-
-}
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-fun ChannelTagsInfo(
-    tagList:List<String>,
-    tagTitle:String,
-    tagTitleLength:Int,
-    changeTagTitle:(String)->Unit,
-    addTag:(String)->Unit,
-    removeTag:(String)->Unit,
-){
-
-    val customTextSelectionColors = TextSelectionColors(
-        handleColor = MaterialTheme.colorScheme.secondary,
-        backgroundColor = MaterialTheme.colorScheme.secondary.copy(alpha = 0.5f)
-    )
-    Column(modifier= Modifier
-        .fillMaxWidth()
-        .padding(horizontal = 10.dp)){
-        Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween){
-            Text("Tags",color = MaterialTheme.colorScheme.onPrimary, fontSize = MaterialTheme.typography.headlineMedium.fontSize)
-            Text("$tagTitleLength",color = MaterialTheme.colorScheme.onPrimary.copy(0.7f))
-        }
-        Divider(thickness = 1.dp, color = MaterialTheme.colorScheme.secondary.copy(0.8f))
-        Spacer(modifier = Modifier.height(5.dp))
-        CompositionLocalProvider(LocalTextSelectionColors provides customTextSelectionColors) {
-
-            TextField(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                singleLine = false,
-                maxLines = 5,
-                value = tagTitle,
-
-                shape = RoundedCornerShape(8.dp),
-                onValueChange = {
-                    val maxLength =24
-                    if (tagTitle.length <= maxLength || it.length < tagTitle.length) {
-                        changeTagTitle(it)
-                    }
-                                },
-                colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
-                    textColor = Color.White,
-                    backgroundColor = Color.DarkGray,
-                    cursorColor = Color.White,
-                    disabledLabelColor = Color.White,
-                    focusedIndicatorColor = Color.Transparent,
-                    unfocusedIndicatorColor = Color.Transparent
-                ),
-                placeholder = {
-                              Text("Enter your own tag",color = Color.White)
-                },
-                trailingIcon = {
-
-                    Icon(
-                        imageVector = Icons.Default.ArrowForward,
-                        contentDescription = "Add new tag",
-                        modifier = Modifier
-                            .size(35.dp)
-                            .clickable {
-                                addTag(tagTitle)
-                            }
-                            .padding(start = 5.dp),
-                        tint = MaterialTheme.colorScheme.onPrimary
-                    )
-
-                }
-            )
-        }
-
-        Spacer(modifier = Modifier.height(10.dp))
-
-            FlowRow(
-                verticalArrangement = Arrangement.spacedBy(10.dp)
-
-            ) {
-                for(item in tagList){
-                    Box(modifier= Modifier
-                        .clip(RoundedCornerShape(5.dp))
-                        .background(Color.DarkGray)
-                        .padding(5.dp)
-                    ){
-                        Row(){
-                            Text(item,color = Color.White)
-                            Icon(
-                                painter = painterResource(id =R.drawable.baseline_close_24),
-                                contentDescription = "Remove tag",
-                                tint = Color.White,
-                                modifier = Modifier.clickable {
-                                    removeTag(item)
-                                }
-                            )
-                        }
-                    }
-                    Spacer(modifier = Modifier.width(5.dp))
-                }
-
-            }
-
-
-    }
-}
-@Composable
-fun ChannelInfoTitle(
-    streamTitle:String,
-    changeStreamTitle:(String)->Unit,
-    titleLength:Int
-){
-    val customTextSelectionColors = TextSelectionColors(
-        handleColor = MaterialTheme.colorScheme.secondary,
-        backgroundColor = MaterialTheme.colorScheme.secondary.copy(alpha = 0.5f)
-    )
-    Column(modifier= Modifier
-        .fillMaxWidth()
-        .padding(horizontal = 10.dp)){
-        Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween){
-            Text("Title",color = MaterialTheme.colorScheme.onPrimary, fontSize = MaterialTheme.typography.headlineMedium.fontSize)
-            Text("$titleLength",color = MaterialTheme.colorScheme.onPrimary.copy(0.7f))
-        }
-        Divider(thickness = 1.dp, color = MaterialTheme.colorScheme.secondary.copy(0.8f))
-        Spacer(modifier = Modifier.height(5.dp))
-        CompositionLocalProvider(LocalTextSelectionColors provides customTextSelectionColors) {
-
-            TextField(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                singleLine = false,
-                maxLines = 5,
-                value = streamTitle,
-
-                shape = RoundedCornerShape(8.dp),
-                onValueChange = {
-                    val maxLength =139
-                    if (streamTitle.length <= maxLength || it.length < streamTitle.length) {
-                        changeStreamTitle(it)
-                    }
-                                },
-                colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
-                    textColor = Color.White,
-                    backgroundColor = Color.DarkGray,
-                    cursorColor = Color.White,
-                    disabledLabelColor = Color.White,
-                    focusedIndicatorColor = Color.Transparent,
-                    unfocusedIndicatorColor = Color.Transparent
-                ),
-                placeholder = {
-                    Text("Enter stream title",color = Color.White)
-                },
-                trailingIcon = {}
-            )
-        }
-
-    }
-}
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
